+++ mutters-bert-intent/README.md
@@ -4,12 +4,12 @@ This is an `IntentMatcher` implementation that uses a Google BERT classification
 
 See https://github.com/google-research/bert for more information on BERT.
 
-Specifically has code for:
+Specifically has Java code for:
 - Reading a BERT model into memory and loading a session
 - Categorizing text using the BERT model (using [TensorFlow Java API](https://github.com/tensorflow/tensorflow/tree/master/tensorflow/java))
 - Adapting the model into an `IntentMatcher` so it can be used with the rest of mutters
 
-To train a model see the [mutters-bert-test-data]() module (TODO!)
+For details on how to train a model see the [mutters-bert-test-data](../mutters-bert-test-data) module (TODO!)
 
 ## Authors
 - [Will Molloy](https://github.com/wilmol)
