+++ mutters-bert-intent/src/main/java/com/rabidgremlin/mutters/bert/doccat/FeatureExtractor.java
@@ -1,10 +1,11 @@
-package com.rabidgremlin.mutters.bert.intent;
+package com.rabidgremlin.mutters.bert.doccat;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Verify.verify;
 
 import java.util.List;
 
+import com.rabidgremlin.mutters.bert.tokenize.Tokenizer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.tensorflow.example.Feature;
@@ -14,7 +15,7 @@ import org.tensorflow.example.Int64List;
 import com.google.common.primitives.Longs;
 
 /**
- * Extracts features from text. Only features are index of each token in the vocab set.
+ * Extracts features from text for document categorization. Only extracts features for tokens in the models vocab.
  * <p>
  * Based on: <a
  * href=https://github.com/google-research/bert/blob/master/extract_features.py>https://github.com/google-research/bert/blob/master/extract_features.py</a>
@@ -28,9 +29,9 @@ public class FeatureExtractor
 
   private final DoccatModel model;
 
-  private final WordPieceTokenizer wordPieceTokenizer;
+  private final Tokenizer wordPieceTokenizer;
 
-  public FeatureExtractor(DoccatModel model, WordPieceTokenizer wordPieceTokenizer)
+  public FeatureExtractor(DoccatModel model, Tokenizer wordPieceTokenizer)
   {
     this.model = checkNotNull(model);
     this.wordPieceTokenizer = checkNotNull(wordPieceTokenizer);
@@ -41,8 +42,8 @@ public class FeatureExtractor
    * <p>
    * Specifically: {@code segment_ids}, {@code label_ids}, {@code input_ids}, and {@code input_mask}.
    * <p>
-   * Note: if the input text (number of words) exceeds the max seq length of the model the result will be truncated on
-   * the right.
+   * Note: if the input text (number of words) exceeds the {@code max_seq_length} of the model the result will be
+   * truncated on the right.
    *
    * @param text text to extract features from
    * @return extracted features
@@ -58,6 +59,7 @@ public class FeatureExtractor
 
     for (int i = 0; i < tokens.size(); i++)
     {
+      // TODO(wilmol) NPE if the token isnt in the vocab
       inputIds[i] = model.vocab().get(tokens.get(i));
       inputMasks[i] = 1;
     }
