+++ mutters-bert-intent/src/main/java/com/rabidgremlin/mutters/bert/intent/BertIntentMatcher.java
@@ -10,6 +10,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.SortedMap;
 
+import com.rabidgremlin.mutters.bert.doccat.DocumentCategorizer;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
